AW9523 must have IO setup priority.
The AW9523 must be set up before its pins are configured so it must have
an earlier setup priority than the GPIO switch and binary sensor
components. Otherwise the pins get set up prematurely and the
configuration gets reset when the AW9523 setup function runs.

Prior to this change, AW9523 had HARDWARE setup priority so it was
ambiguous whether it would be set up before or after GPIO components
which also have HARDWARE priority. Indeed, this happened in my
configuration and it did not work correctly.